+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/option/SoraSpotlightOption.kt
@@ -1,6 +1,6 @@
 package jp.shiguredo.sora.sdk.channel.option
 
-import jp.shiguredo.sora.sdk.channel.signaling.message.SimulcastRid
+import jp.shiguredo.sora.sdk.channel.option.SoraSimulcastOption
 
 /**
  * スポットライト機能のオプションです。
@@ -10,7 +10,7 @@ class SoraSpotlightOption {
     /**
      * サイマルキャストの rid
      */
-    var simulcastRid: SimulcastRid? = null
+    var simulcastRid: SoraSimulcastOption.SimulcastRid? = null
 
     /**
      * スポットライト機能のアクティブな配信数を指定します