+++ CHANGES.md
@@ -11,9 +11,7 @@
 
 ## develop
 
-- [CHANGE] ConnectMessage.channelId の型を String? から String に変更する
-  - @enm10k
-- [UPDATE] NotificationMessage に turnTransportType を追加する
+- [CHANGE] SoraMediaChannel のコンストラクタ引数 channelId の型を String? から String に変更する
   - @enm10k
 - [ADD] データチャネルシグナリングに対応する
   - data_channel_signlaing, ignore_disconnect_websocket パラメータ設定を追加する
@@ -26,6 +24,8 @@
   - @shino
 - [FIX] offer に data_channels が含まれない場合に対応する
   - @shino
+- [FIX] NotificationMessage に turnTransportType を追加する
+  - @enm10k
 
 ## 2021.1